+++ src/core/managers/UnlockPanelManager/UnlockPanelManager.types.ts
@@ -17,11 +17,6 @@ export enum UnlockPanelEventsEnum {
   ANCHOR_CLOSE = 'ANCHOR_CLOSE'
 }
 
-export interface IUnlockPanel {
-  isOpen: boolean;
-  allowedProviders?: IProviderBase[] | null;
-}
-
 export type LoginFunctonType = ({
   type,
   anchor
@@ -36,6 +31,7 @@ export type CustomProviderViewType<
 export type LoginCallbackType = () => void;
 
 export type LoginHandlerType = LoginFunctonType | LoginCallbackType;
+export type CloseCallbackType = () => void;
 
 export type UnlockPanelManagerInitParamsType = {
   /**
@@ -63,4 +59,12 @@ export type UnlockPanelManagerInitParamsType = {
    * List of allowed providers
    */
   allowedProviders?: AllowedProviderType[] | null;
+  /**
+   * Callback function to handle UI behavior when the unlock panel is closed
+   * without completing the login process.
+   *
+   * Common use case: redirecting the user away from the `/unlock` route
+   * (e.g., back to the homepage or a previous screen) when login is cancelled or skipped.
+   */
+  closeCallback?: CloseCallbackType;
 };