+++ CHANGELOG.md
@@ -7,12 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Update unlock panel to use side panel base](https://github.com/multiversx/mx-sdk-dapp/pull/1427)
+- [Refactor and streamline side panel managers, event handling, and state management for transaction-related UI flows.](https://github.com/multiversx/mx-sdk-dapp/pull/1426)
+- [Add closeCallback in UnlockPanelManager](https://github.com/multiversx/mx-sdk-dapp/pull/1425)
+
 ## [[5.0.0-alpha.0](https://github.com/multiversx/mx-sdk-dapp/pull/1421)] - 2025-05-20
 
 - [Edit project configs](https://github.com/multiversx/mx-sdk-dapp/pull/1423)
 - [Rename to sdk-dapp](https://github.com/multiversx/mx-sdk-dapp/pull/1420)
 
-----------------
+---
 
 - [Upgrade sdk-dapp-ui](https://github.com/multiversx/mx-sdk-dapp-core/pull/172)
 - [Fix walletConnect login issues](https://github.com/multiversx/mx-sdk-dapp-core/pull/171)
@@ -173,7 +177,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[v0.0.0]](https://github.com/multiversx/mx-sdk-dapp-core)] - 2024-04-17
 
----------------------
+---
 
 ## [[v4.1.8](https://github.com/multiversx/mx-sdk-dapp/pull/1418)] - 2025-05-20